requirements/base.txt
celery==5.2.3
Django==4.0.1
django-extensions==3.1.5
django-model-utils==3.1.2
django-rest-auth==0.9.5
djangorestframework==3.13.1
ipython==8.0.1
isort==5.10.1
Pillow==9.0.0
psycopg2==2.9.3
python-gettext==4.0
pytz==2021.3
django-filter==21.1
factory_boy==3.2.1
model_mommy==2.0.0
moviepy==1.0.3